requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
azure-core==1.32.0
azure-cosmos==4.3.1
azure-storage-blob==12.24.1
bcrypt==4.2.1
blinker==1.9.0
cachetools==5.5.2
captcha==0.7.0
certifi==2022.12.7
cffi==1.17.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==44.0.1
distlib==0.3.6
distro==1.9.0
extra-streamlit-components==0.1.71
filelock==3.12.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.4
isodate==0.7.2
Jinja2==3.1.5
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.27.1
numpy==2.2.3
openai==1.64.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pip-upgrade==0.0.6
platformdirs==3.2.0
protobuf==5.29.3
pur==7.1.0
pyarrow==19.0.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.28.2
rich==13.9.4
rpds-py==0.23.1
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.42.2
streamlit-authenticator==0.4.1
streamlit-cookies-manager==0.2.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.15
virtualenv==20.22.0
virtualenv-clone==0.5.7
watchdog==6.0.0
